Allow signing repodata while exporting products
Resolves: https://github.com/AlmaLinux/build-system/issues/444
