+++ scripts/exporters/products_exporter.py
@@ -1,5 +1,6 @@
 import argparse
 import asyncio
+import os
 import sys
 from pathlib import Path
 from typing import List, Literal
@@ -65,6 +66,13 @@ def parse_args():
         default="hardlink",
         required=False,
         help="Method of exporting (choices: write, hardlink, symlink)",
+    )
+    parser.add_argument(
+        "-sw",
+        "--sign-with",
+        type=str,
+        required=True,
+        help="GPG key name to use when signing repodata.",
     )
     parser.add_argument(
         "-l",
@@ -137,6 +145,17 @@ class ProductExporter(BasePulpExporter):
             list({repo.id for repo in product.repositories})
         )
 
+    async def get_sign_key_id(self, key_id):
+        sign_keys = await self.get_sign_keys()
+        return next(
+            (
+                sign_key['keyid']
+                for sign_key in sign_keys
+                if sign_key['keyid'] == key_id
+            ),
+            None,
+        )
+
 
 async def repo_post_processing(
     exporter: ProductExporter,
@@ -152,6 +171,27 @@ async def repo_post_processing(
         return result
 
 
+async def sign_repodata(
+    exporter: ProductExporter,
+    exported_paths: List[str],
+    key_id: str,
+):
+    tasks = []
+    token = await exporter.get_sign_server_token()
+
+    for repo_path in exported_paths:
+        path = Path(repo_path)
+        parent_dir = path.parent
+        repodata = parent_dir / "repodata"
+        if not os.path.exists(repo_path):
+            continue
+
+        exporter.logger.info('Key ID: %s', str(key_id))
+        tasks.append(exporter.repomd_signer(repodata, key_id, token))
+
+    await asyncio.gather(*tasks)
+
+
 async def main():
     args = parse_args()
     await setup_all()
@@ -161,6 +201,14 @@ async def main():
         export_method=args.export_method,
         log_file_path=args.log,
     )
+
+    sign_key_id = None
+    if args.sign_with:
+        sign_key_id = await exporter.get_sign_key_id(args.sign_with)
+        if not sign_key_id:
+            err = "Couldn't retrieve the '{args.sign_with}' sign key"
+            raise Exception(f'Aborting product export, error was: {err}')
+
     pulp_client.PULP_SEMAPHORE = asyncio.Semaphore(10)
     exported_paths = await exporter.export_product_repos(
         product_name=args.product,
@@ -170,6 +218,9 @@ async def main():
     await asyncio.gather(
         *(repo_post_processing(exporter, path) for path in exported_paths)
     )
+
+    if sign_key_id:
+        await sign_repodata(exporter, exported_paths, sign_key_id)
 
 
 if __name__ == '__main__':
